+++ inst/extdata/try1.R
@@ -1,0 +1,1 @@
+dat <- read.table("inst/extdata/intcal20_data.txt", sep="\t")