+++ NEWS.md
@@ -1,3 +1,7 @@
+# IntCal 0.2.0
+* added a function intcal.data to plot the data underlying the IntCal curves.
+* to add another function to quickly plot the curves (1 or 2, or to add a curve to an existing plot)
+
 # IntCal 0.1.3
 * fixed issues with the DESCRIPTION file